+++ src/Sulu/Bundle/SecurityBundle/SuluSecurityBundle.php
@@ -11,6 +11,7 @@
 namespace Sulu\Bundle\SecurityBundle;
 
 use Sulu\Bundle\SecurityBundle\DependencyInjection\Compiler\CurrentUserDataCompilerPass;
+use Sulu\Bundle\SecurityBundle\DependencyInjection\Compiler\SecurityCheckerCompilerPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
@@ -21,5 +22,6 @@ class SuluSecurityBundle extends Bundle
         parent::build($container);
 
         $container->addCompilerPass(new CurrentUserDataCompilerPass());
+        $container->addCompilerPass(new SecurityCheckerCompilerPass());
     }
 }